+++ src/main/java/com/muud/diary/controller/DiaryController.java
@@ -1,13 +1,13 @@
-package com.muud.diary.presentation;
+package com.muud.diary.controller;
 
 import com.muud.auth.jwt.Auth;
-import com.muud.diary.application.DiaryService;
+import com.muud.diary.service.DiaryService;
 import com.muud.diary.domain.Diary;
-import com.muud.diary.dto.ContentUpdateRequest;
-import com.muud.diary.dto.DiaryPreviewResponse;
-import com.muud.diary.dto.DiaryRequest;
-import com.muud.diary.dto.DiaryResponse;
-import com.muud.emotion.entity.Emotion;
+import com.muud.diary.domain.dto.ContentUpdateRequest;
+import com.muud.diary.domain.dto.DiaryPreviewResponse;
+import com.muud.diary.domain.dto.DiaryRequest;
+import com.muud.diary.domain.dto.DiaryResponse;
+import com.muud.emotion.domain.Emotion;
 import com.muud.user.entity.User;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
