+++ src/main/java/com/muud/diary/domain/dto/ContentUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.muud.diary.dto;
+package com.muud.diary.domain.dto;
 
 import jakarta.validation.constraints.Size;
