+++ src/main/java/com/muud/emotion/domain/dto/EmotionCountResponse.java
@@ -1,4 +1,4 @@
-package com.muud.emotion.dto;
+package com.muud.emotion.domain.dto;
 
 public record EmotionCountResponse(String emotionName,
                                    Long count) {
