+++ src/main/java/com/muud/emotion/domain/dto/QuestionResponse.java
@@ -1,4 +1,4 @@
-package com.muud.emotion.dto;
+package com.muud.emotion.domain.dto;
 
 public record QuestionResponse(String emotionName,
                                String question) {
