+++ src/main/java/com/muud/playlist/repository/PlayListRepository.java
@@ -1,13 +1,12 @@
 package com.muud.playlist.repository;
 
-import com.muud.emotion.entity.Emotion;
+import com.muud.emotion.domain.Emotion;
 import com.muud.playlist.entity.PlayList;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface PlayListRepository extends JpaRepository<PlayList, Long> {
