+++ src/main/java/com/muud/report/domain/dto/RankDto.java
@@ -1,4 +1,4 @@
-package com.muud.report.dto;
+package com.muud.report.domain.dto;
 
 public record RankDto<T>(int rank,
                       T object) {
